requirements.txt
django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg2==2.9.9
python-dotenv==1.0.1
whitenoise==6.6.0
gunicorn==22.0.0
dj-rest-auth==6.0.0
django-allauth==0.62.1
drf-spectacular==0.27.2
requests==2.31.0
djangorestframework-simplejwt==5.2.2
django-filter==24.2

